NumberVerifier 5.0.1.0 - SDLCOM-6445:
- Resolved issue where different values appeared in different views
- Fixed UI inconsistencies
- Corrected error severity for QAs
- Fixed issue with settings being lost
- Ensured Number Verifier QA is performed correctly
- Added import/export functionality for settings (XML files)
- Added Helper message when the current settings mismatch the current profile
